refactor: extract /productivity routes into productivity-router (#585)
Pure refactor — no logic changes. Splits the productivity routes (six
endpoints) out of activities-router.ts into a dedicated
`routes/productivity-router.ts`, dropping activities-router from 788 → 655
lines as suggested in #585.

The new router is mounted at `/productivity` (paths inside the router are
relative). Activities-router only handles `/activities/*` now.

- routes/activities-router.ts: 788 → 655 lines
- routes/productivity-router.ts: 148 lines (new)
