+++ apps/backend/src/api/rest-routes.ts
@@ -31,6 +31,7 @@ import { createLocationsRouter } from '../routes/locations-router.ts'
 import { createMealsRouter } from '../routes/meals-router.ts'
 import { createMetricsRouter } from '../routes/metrics-router.ts'
 import { createNotesRouter } from '../routes/notes-router.ts'
+import { createProductivityRouter } from '../routes/productivity-router.ts'
 import { createRawRecordsRouter } from '../routes/raw-records-router.ts'
 import { createReportsRouter } from '../routes/reports-router.ts'
 import { createScreentimeCategoriesRouter } from '../routes/screentime-categories-router.ts'
@@ -86,6 +87,7 @@ export const mountRestRouters = ({
       },
     ),
   )
+  httpd.use('/productivity', createProductivityRouter(authMiddleware, syncProvider))
   httpd.use('/activity-types', createActivityTypesRouter(authMiddleware))
   httpd.use(
     '/deduction-rules',
